services/logger/requirements.txt
flask==3.*
gunicorn==23.*
google-cloud-firestore==2.*
